+++ tests/batch_group_tests/batch_group_creation_interface_2.py
@@ -6,14 +6,12 @@ def adversary_create(batch_job):
     adversary = batch_job.group('Adversary', 'adversary-002')
     adversary.attribute('Description', 'Example Description', True)
     adversary.tag('Example Tag')
-    adversary.security_label('TLP Green')
 
 
 def campaign_create(batch_job):
     campaign = batch_job.group('Campaign', 'campaign-002')
     campaign.attribute('Description', 'Example Description', True)
     campaign.tag('Example Tag')
-    campaign.security_label('TLP Green')
 
 
 def document_create(batch_job):
@@ -21,7 +19,6 @@ def document_create(batch_job):
     document.add_file('test.txt', 'Document content')
     document.attribute('Description', 'Example Description', True)
     document.tag('Example Tag')
-    document.security_label('TLP Green')
 
 
 def document_malware_create(batch_job):
@@ -35,7 +32,6 @@ def document_malware_create(batch_job):
     document.add_key_value('password', 'test')
     document.attribute('Description', 'Example Description', True)
     document.tag('Example Tag')
-    document.security_label('TLP Green')
 
 
 def email_create(batch_job):
@@ -45,28 +41,24 @@ def email_create(batch_job):
     email.add_key_value('body', 'This is just a test')
     email.attribute('Description', 'Example Description', True)
     email.tag('Example Tag')
-    email.security_label('TLP Green')
 
 
 def event_create(batch_job):
     event = batch_job.group('Event', 'event-002')
     event.attribute('Description', 'Example Description', True)
     event.tag('Example Tag')
-    event.security_label('TLP Green')
 
 
 def incident_create(batch_job):
     incident = batch_job.group('Incident', 'incident-002')
     incident.attribute('Description', 'Example Description', True)
     incident.tag('Example Tag')
-    incident.security_label('TLP Green')
 
 
 def intrusion_set_create(batch_job):
     intrusion_set = batch_job.group('Intrusion Set', 'intrusion_set-002')
     intrusion_set.attribute('Description', 'Example Description', True)
     intrusion_set.tag('Example Tag')
-    intrusion_set.security_label('TLP Green')
 
 
 def report_create(batch_job):
@@ -74,7 +66,6 @@ def report_create(batch_job):
     report.add_file('test.txt', 'Report content')
     report.attribute('Description', 'Example Description', True)
     report.tag('Example Tag')
-    report.security_label('TLP Green')
 
 
 def signature_create(batch_job):
@@ -84,11 +75,9 @@ def signature_create(batch_job):
     signature.add_key_value('fileText', 'This is just a test...')
     signature.attribute('Description', 'Example Description', True)
     signature.tag('Example Tag')
-    signature.security_label('TLP Green')
 
 
 def threat_create(batch_job):
     threat = batch_job.group('Threat', 'threat-002')
     threat.attribute('Description', 'Example Description', True)
     threat.tag('Example Tag')
-    threat.security_label('TLP Green')