+++ tests/batch_indicator_tests/batch_ind_creation_interface_1.py
@@ -10,7 +10,7 @@ def address_create(batch_job):
 
 
 def email_address_create(batch_job):
-    email_address = batch_job.email_address('bad@example.com', '5.0', '100')
+    email_address = batch_job.email_address('bad@dfadsfaddsfa.com', '5.0', '100')
     email_address.attribute('Description', 'Example Description', True)
     email_address.tag('Example Tag')
     email_address.security_label('TLP Green')
@@ -24,14 +24,14 @@ def file_create(batch_job):
 
 
 def host_create(batch_job):
-    host = batch_job.host('example.com', rating='5.0', confidence='100')
+    host = batch_job.host('dfadsfaddsfa.com', rating='5.0', confidence='100')
     host.attribute('Description', 'Example Description', True)
     host.tag('Example Tag')
     host.security_label('TLP Green')
 
 
 def url_create(batch_job):
-    url = batch_job.url('https://example.com/index.html', '5.0', '100')
+    url = batch_job.url('https://dfadsfaddsfa.com/index.html', '5.0', '100')
     url.attribute('Description', 'Example Description', True)
     url.tag('Example Tag')
     url.security_label('TLP Green')